Add an Event system
https://github.com/robertwahler/EventManager

I'm fond of this Event system and use it heavily for controlling game flow. These files include some very minor edits by myself, which are on Github as a fork.

I really wanted to add this as a Git subtree. However, I only want a small handful of the repo's files and I'm not knowledgable enough with Git to figure out the Split command stuff. Perhaps another day.