Add Unity's Localization package (preview)
Text is a major element of DQU, but before I add any, I figure it'd be best to adopt a Localization pattern/system. As luck would have it, Unity has a Localization package in the works. It's still early and not very user-friendly, but let's roll with it for now.