Add player character
Working on art is already a slow process for me these days, and when life intrudes in a big way? Apparently the result is months without any commits. Youch.

Anyway, the player now controls an actual animated character now. It uses a 3D model that I've hopefully made to resemble a sprite.
   • Some basic support for different clothing/armor.
   • Shader support for "palette"-based coloring.
   • Quite a lot of Editor GUI work to make this stuff easy to work with.